refactor: Refactor the full text index
Functionally this is the same, except for logging.
But see what's coming next...
