+++ packages/cli/package.json
@@ -60,7 +60,7 @@
     "eslint-plugin-prettier": "^3.4.0",
     "http-browserify": "^1.7.0",
     "https-browserify": "^1.0.0",
-    "jest": "^29.6.3",
+    "jest": "^29.7.0",
     "jest-sinon": "^1.1.0",
     "lint-staged": "^10.5.4",
     "memfs": "^3.4.13",
